Remove unused types and functions. Implement documentation for new functions.
