+++ gin-server/server.go
@@ -58,36 +58,36 @@ func saveNewAccount(c *gin.Context) {
 	c.JSON(http.StatusOK, gin.H{})
 }
 
-// retrieveAccounts
-// returns value of GetAllAccount()
-func retrieveAccounts(c *gin.Context) {
-	c.JSON(http.StatusOK, gin.H{
-		"accounts": db_accounts.GetAllAccount(),
-	})
-}
-
+// getAccountNames
+// returns all account name from db
+// - returns 200 and array of names if successful
+// - returns error message if any error occurs
 func getAccountNames(c *gin.Context) {
 	c.JSON(http.StatusOK, gin.H{
 		"names": db_accounts.GetAccountNames(),
 	})
 }
 
+// getAccount
+// returns account based on provided account name
+// - returns 200 and array of names if successful
+// - returns error message if any error occurs
 func getAccount(c *gin.Context) {
 	type Payload struct {
 		AccName string `json:"name"`
 	}
 	var payload Payload
-	//
+	// prompt if error occurs during data retrieval
 	if err := c.BindJSON(&payload); err != nil {
 		c.JSON(400, gin.H{"error": err.Error()})
 	}
-	//
 	acc, _ := db_accounts.GetAccount(payload.AccName)
 	c.JSON(http.StatusOK, gin.H{
 		"account": acc,
 	})
 }
 
+// main
 // contains endpoint initialization and handlers
 func main() {
 	// initialize gin
@@ -96,7 +96,6 @@ func main() {
 	router.POST("/upload", rawUpload)
 	router.POST("/saveNewAccount", saveNewAccount)
 	router.POST("/getAccount", getAccount)
-	router.GET("/retrieveAccounts", retrieveAccounts)
 	router.GET("/getAccountNames", getAccountNames)
 	// run via localhost:5000
 	err := router.Run(":5000")
