+++ gin-server/types/account.go
@@ -1,6 +1,8 @@
 package types
 
 type Account struct {
-	AccName string `json:"AccName"`
-	Data    AccountData
+	AccName string
+	RowData []map[string]string
+	ColDefs []ColDef
+	TagDefs []TagDef
 }
