--- gin-server/types/accountData.go
@@ -1,7 +1,0 @@
-package types
-
-type AccountData struct {
-	RowData []map[string]string
-	ColDefs []ColDef
-	TagDefs []TagDef
-}
