--- gin-server/types/accs.go
@@ -1,8 +1,0 @@
-package types
-
-type Accs struct {
-	AccName string
-	RowData []map[string]string
-	ColDefs []ColDef
-	TagDefs []TagDef
-}
