+++ src/Types/IAccount.ts
@@ -1,7 +1,8 @@
 import {Row} from "./Row";
 import {ColDef} from "ag-grid-community";
 
-export interface IAccountData {
+export interface IAccount {
+    AccName: string,
     RowData: Row[],
     ColDefs: ColDef[],
     TagDefs: ColDef[],
