--- src/Types/IAccount.tsx
@@ -1,6 +1,0 @@
-import {IAccountData} from "./IAccountData";
-
-export interface IAccount {
-    AccName: string,
-    Data: IAccountData
-}
